drop unused method
